Fix compilation errors in firmware

+++ board/econet.h
@@ -63,7 +63,7 @@ bool                    econet_init(
                             uint8_t*    ack_buffer,
                             size_t      ack_buffer_sz);
 econet_tx_result_t      broadcast(uint8_t* buff, int bytes);
-econet_tx_result_t      transmit(uint8_t station, uint8_t network, uint8_t control, uint8_t port, uint* data, size_t data_len);
+econet_tx_result_t      transmit(uint8_t station, uint8_t network, uint8_t control, uint8_t port, uint8_t* data, size_t data_len);
 econet_rx_result_t      receive();
 econet_rx_result_t      monitor();
 uint8_t                 get_station();
